finished creating seed data for all tables

+++ backend/db/migrations/20220923223146-create-review.js
@@ -19,7 +19,7 @@ module.exports = {
         allowNull: false
       },
       stars: {
-        type: Sequelize.INTEGER
+        type: Sequelize.INTEGER(1)
       },
       createdAt: {
         allowNull: false,
